% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base test for 3.R
\name{anof}
\alias{anof}
\title{Perform Analysis of Variance with Filtering}
\usage{
anof(data, variable, id, formula, ...)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{variable}{A character string specifying the variable in the dataset used for filtering.}

\item{id}{A character string specifying the identifier value to filter the data.}

\item{formula}{A formula specifying the model for the analysis of variance.}

\item{...}{Additional arguments to be passed to the \code{\link[stats]{aov}} function.}
}
\value{
An ANOVA object.
}
\description{
This function conducts an analysis of variance (ANOVA) after filtering data based on a specified variable and its corresponding identifier.
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")

df <- ToothGrowth
anof(data = df, variable = "supp", id = "VC", formula = len ~ dose)
}

}
\seealso{
\code{\link[stats]{aov}} for more advanced t-test functionality.
}
