% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get p value 3.R
\name{dunn_p}
\alias{dunn_p}
\title{Perform Post-hoc Test using Dunn's Test}
\usage{
dunn_p(data, group, variable, id, formula, method = "bonferroni", ...)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group}{A character string specifying the grouping variable in the dataset.}

\item{variable}{A character string specifying the variable in the dataset used for filtering.}

\item{id}{A character string specifying the identifier value to filter the data.}

\item{formula}{A formula specifying the model for the post-hoc test. For example, formula = TP53 ~ cancer_group.}

\item{method}{Allowed values include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default is "bonferroni".}

\item{...}{Additional arguments to be passed.}
}
\value{
A data frame containing the p-values and additional information for each comparison.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function conducts a post-hoc test using Dunn's test with a
specified adjustment method after filtering data based on a specified
variable and its corresponding identifier.
}\if{html}{\out{</div>}}
}
\examples{
\donttest{
# Load data
data("ToothGrowth")

df <- ToothGrowth
dunn_p(data = df, group = "dose", variable = "supp", id = "VC", formula = len ~ dose)
}

}
\seealso{
\code{\link[rstatix]{dunn_test}} for more information on Dunn's test.
}
