% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base test.R
\name{format_p}
\alias{format_p}
\title{Format P-values for display}
\usage{
format_p(p, small_thresh = 1e-04, digits = 4)
}
\arguments{
\item{p}{A numeric p-value.}

\item{small_thresh}{Threshold below which the p-value will be displayed
in scientific notation. Defaults to \code{1e-4}.}
}
\value{
A character string representing the formatted p-value.
}
\description{
This function formats numeric p-values for reporting.
Values smaller than \code{small_thresh} are displayed in scientific notation
with 2 significant digits. Other values are rounded to four decimal places.
}
\examples{
format_p(0.123456)
format_p(0.0000002)
format_p(NA)
}
