% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get p value 3.R
\name{hsd_p}
\alias{hsd_p}
\title{Perform Post-hoc Test for Tukey's HSD}
\usage{
hsd_p(data, group, variable, id, formula, method = "hsd", ...)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group}{A character string specifying the grouping variable in the dataset.}

\item{variable}{A character string specifying the variable in the dataset used for filtering.}

\item{id}{A character string specifying the identifier value to filter the data.}

\item{formula}{A formula specifying the model for the analysis of variance.}

\item{method}{one of "hsd", "bonf", "lsd", "scheffe", "newmankeuls", default is "hsd".}

\item{...}{Additional arguments to be passed.}
}
\value{
A data frame containing the adjusted p-values and additional information for each comparison.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function conducts a post-hoc test using Tukey's Honestly Significant
Difference (HSD) method after conducting an ANOVA and filtering data
based on a specified variable and its corresponding identifier.
}\if{html}{\out{</div>}}
}
\examples{
\donttest{
# Load data
data("ToothGrowth")

df <- ToothGrowth
hsd_p(data = df, group = "dose", variable = "supp", id = "VC", formula = len ~ dose)
}

}
