% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base test for 3.R
\name{kwf}
\alias{kwf}
\title{Perform Kruskal-Wallis Test with Filtering}
\usage{
kwf(data, variable, id, formula, ...)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{variable}{A character string specifying the variable in the dataset used for filtering.}

\item{id}{A character string specifying the identifier value to filter the data.}

\item{formula}{A formula specifying the model for the Kruskal-Wallis test.}

\item{...}{Additional arguments to be passed to the \code{\link[rstatix]{kruskal_test}} function.}
}
\value{
A Kruskal-Wallis test object.
}
\description{
This function conducts a Kruskal-Wallis test after filtering data based on a specified variable and its corresponding identifier.

This function conducts a Kruskal-Wallis test after filtering data based on a specified variable and its corresponding identifier.
}
\examples{
\donttest{
# Load data
data("ToothGrowth")

df <- ToothGrowth
kwf(data = df, variable = "supp", id = "VC", formula = len ~ dose)
}
}
\seealso{
\code{\link[rstatix]{kruskal_test}} for more information on Kruskal-Wallis test functionality.
}
