% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get p value 3.R
\name{lsd_p}
\alias{lsd_p}
\title{Perform Post-hoc Test for Least Significant Difference (LSD)}
\usage{
lsd_p(data, group, variable, id, formula, method = "lsd", ...)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group}{A character string specifying the grouping variable in the dataset.}

\item{variable}{A character string specifying the variable in the dataset used for filtering.}

\item{id}{A character string specifying the identifier value to filter the data.}

\item{formula}{A formula specifying the model for the analysis of variance.}

\item{method}{one of "hsd", "bonf", "lsd", "scheffe", "newmankeuls", defining the method for the pairwise comparisons, default is "lsd".}

\item{...}{Additional arguments to be passed.}
}
\value{
A data frame containing the p-values and additional information for each comparison.
}
\description{
This function conducts a post-hoc test using the Least Significant Difference (LSD) method after conducting an ANOVA and filtering data based on a specified variable and its corresponding identifier.
}
\examples{
\donttest{
# Load data
data("ToothGrowth")

df <- ToothGrowth
lsd_p(data = df, group = "dose", variable = "supp", id = "VC", formula = len ~ dose)
}

}
\seealso{
\code{\link[DescTools]{PostHocTest}} for more information on post-hoc tests.
}
