% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base test.R
\name{welch_tf}
\alias{welch_tf}
\title{Welch's T-test}
\usage{
welch_tf(data, variable, id, formula, var.equal = FALSE, ...)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the t-test will be conducted.}

\item{id}{A value to filter the \code{variable} by.}

\item{formula}{An optional formula specifying the structure of the statistical model to be tested.}

\item{var.equal}{Logical indicating whether to assume equal variances in the two groups. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{\link[rstatix]{t_test}}.}
}
\value{
A t-test result.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Conducts a Welch's t-test on a specified variable in the provided
data.
}\if{html}{\out{</div>}}
}
\details{
This function filters the data based on the provided \code{variable}
and \code{id}, and then conducts a welch's t-test on the filtered
subset of data. If a formula is provided, the t-test is conducted based
on the specified model structure.

This function filters the data based on the provided \code{variable} and \code{id}, and then conducts a t-test on the filtered subset of data. If a formula is provided, the t-test is conducted based on the specified model structure.
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth\%>\%
filter(dose \%in\% c("0.5","1"))

welch_tf(data = df,variable = "supp",id="VC",formula = len ~ dose)
}
}
\seealso{
\code{\link[rstatix]{t_test}} for more advanced t-test functionality.
}
