% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get p value.R
\name{weltf_p}
\alias{weltf_p}
\title{For P Value (Welch's t-test)}
\usage{
weltf_p(id, ...)
}
\arguments{
\item{id}{A value to filter the \code{variable} by.}

\item{...}{Additional arguments to be passed to \code{\link[rstatix]{t_test}}.}

\item{data}{A data frame containing the data.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the t-test will be conducted.}

\item{formula}{An optional formula specifying the structure of the statistical model to be tested.}

\item{var.equal}{Logical indicating whether to assume equal variances in the two groups. Default is \code{FALSE}.}
}
\value{
A data frame containing 'group1', 'group2', 'p', 'p.adj', and
'variable'.
}
\description{
For P Value (Welch's t-test)
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth\%>\%
filter(dose \%in\% c("0.5","1"))

weltf_p(data = df,variable = "supp",id="VC",formula = len ~ dose)
}
}
