% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base test.R
\name{wf}
\alias{wf}
\title{Wilcoxon Tests}
\usage{
wf(data, variable, id, formula, var.equal = FALSE, ...)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the wilcox test will be conducted.}

\item{id}{A value to filter the \code{variable} by.}

\item{formula}{An optional formula specifying the structure of the statistical model to be tested.}

\item{var.equal}{Logical indicating whether to assume equal variances in the two groups. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{\link[rstatix]{wilcox_test}}.}
}
\value{
A wilcoxon test result.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Conducts a Wilcoxon test on a specified variable in the provided
data.
}\if{html}{\out{</div>}}
}
\details{
This function filters the data based on the provided \code{variable}
and \code{id}, and then conducts a Wilcoxon test on the filtered
subset of data. If a formula is provided, the Wilcoxon is conducted based
on the specified model structure.

This function filters the data based on the provided \code{variable} and \code{id}, and then conducts a wilcox test on the filtered subset of data. If a formula is provided, the wilcox test is conducted based on the specified model structure.
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth\%>\%
filter(dose \%in\% c("0.5","1"))

wf(data = df,variable = "supp",id="VC",formula = len ~ dose)
}
}
\seealso{
\code{\link[rstatix]{wilcox_test}} for more advanced wilcoxon test functionality.
}
