% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get p value.R
\name{wf_p}
\alias{wf_p}
\title{For P Value (Wilcoxon test)}
\usage{
wf_p(id, ...)
}
\arguments{
\item{id}{A value to filter the \code{variable} by.}

\item{...}{Additional arguments to be passed to \code{\link[rstatix]{wilcox_test}}.}

\item{data}{A data frame containing the data.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the wilcox_test will be conducted.}

\item{formula}{An optional formula specifying the structure of the statistical model to be tested.}
}
\value{
A data frame containing 'group1', 'group2', 'p', 'p.adj', and
'variable'.
}
\description{
For P Value (Wilcoxon test)
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth\%>\%
filter(dose \%in\% c("0.5","1"))

wf_p(data = df,variable = "supp",id="VC",formula = len ~ dose)
}
}
