% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two groups comparision.R
\name{stat2}
\alias{stat2}
\title{Perform statistical tests for two-group comparison}
\usage{
stat2(
  data,
  group,
  value,
  formula,
  variable = FALSE,
  id = "id",
  sap.size = 30,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{group}{The grouping variable for comparing groups.}

\item{value}{The variable representing the values to be analyzed.}

\item{formula}{A formula specifying the relationship between variables.An optional formula specifying the structure of the statistical model to be tested.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the test will be conducted. Ignore if not available.}

\item{id}{The identifier for each observation.A value to filter the \code{variable} by. Ignore if not available.}

\item{sap.size}{The sample size threshold for determining whether to use parametric tests. Default is 30.}

\item{...}{Additional arguments to be passed to the statistical tests.}
}
\value{
A list containing the results of the statistical tests.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Depending on the characteristics of the data, it automatically determines whether to use parametric tests (t-test, Welch's t-test) or non-parametric tests (Wilcoxon test).
}\if{html}{\out{</div>}}
}
\examples{
\donttest{
# Load data
#:::::::::::::::::::::::::::::::::::::::
# Dataframe with multiple columns, need to filter variable
data("ToothGrowth")

df <- ToothGrowth\%>\%
filter(dose \%in\% c("0.5","1"))

stat2(data = df,variable = "supp",id="VC",group = "dose",value = "len",
formula = len ~ dose)

#:::::::::::::::::::::::::::::::::::::::
# Dataframe with only two columns (group,value)
data("HairEyeColor")
df <- as.data.frame(HairEyeColor)[,c(3,4)]

stat2(data = df,group = "Sex",value = "Freq", formula = Freq ~ Sex) # Ignoring variable and id
}
}
