% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more than two groups comparision.R
\name{stat3}
\alias{stat3}
\title{Perform Statistical Test for More Than Two-group Comparison}
\usage{
stat3(data, group, value, formula, variable = FALSE, id = "id", ...)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{group}{The grouping variable for comparing groups.}

\item{value}{The variable representing the values to be analyzed.}

\item{formula}{A formula specifying the relationship between variables.An optional formula specifying the structure of the statistical model to be tested.}

\item{variable}{A character string specifying the name of the variable in \code{data} for which the test will be conducted. Ignore if not available.}

\item{id}{The identifier for each observation.A value to filter the \code{variable} by. Ignore if not available.}

\item{...}{Additional arguments to be passed to the statistical tests.}
}
\value{
A data frame containing the results of the statistical test.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Depending on the characteristics of the data, it automatically determines
whether to use parametric tests (Anova, HSD(post-hoc)) or non-parametric
tests (kruskal, Dunn's(post-hoc)).
}\if{html}{\out{</div>}}
}
\examples{
\donttest{
# Load data
# Dataframe with multiple columns, need to filter variable
data("ToothGrowth")

df <- ToothGrowth
stat3(data = df, group = "dose", value = "len", variable = "supp", id = "OJ", formula = len ~ dose)

#:::::::::::::::::::::::::::::::::::::::
# Dataframe with only two columns (group,value)
data("HairEyeColor")
df <- as.data.frame(HairEyeColor)[,c(2,4)]

stat3(data = df,group = "Eye",value = "Freq", formula = Freq ~ Eye) # Ignoring variable and id
}

}
